+++ src/lib/builders/select/index.ts
@@ -8,11 +8,13 @@ import {
 	getElementByMeltId,
 	isBrowser,
 	kbd,
+	omit,
 	styleToString,
 	uuid,
 } from '$lib/internal/helpers';
 import { sleep } from '$lib/internal/helpers/sleep';
 import type { Defaults } from '$lib/internal/types';
+import { tick } from 'svelte';
 import { derived, writable } from 'svelte/store';
 
 /**
@@ -31,6 +33,8 @@ export type CreateSelectArgs = {
 	arrowSize?: number;
 	required?: boolean;
 	disabled?: boolean;
+	selected?: string | number;
+	name?: string;
 };
 
 const defaults = {
@@ -45,11 +49,11 @@ const defaults = {
 
 export function createSelect(args?: CreateSelectArgs) {
 	const withDefaults = { ...defaults, ...args } as CreateSelectArgs;
-	const options = writable({ ...withDefaults });
+	const options = writable(omit(withDefaults, 'selected'));
 
 	const open = writable(false);
-	const selected = writable<string | null>(null);
-	const selectedText = writable<string | null>(null);
+	const selected = writable(withDefaults.selected ?? null);
+	const selectedText = writable<string | number | null>(null);
 	const activeTrigger = writable<HTMLElement | null>(null);
 
 	const ids = {
@@ -75,6 +79,7 @@ export function createSelect(args?: CreateSelectArgs) {
 				style: styleToString({
 					display: $open ? undefined : 'none',
 				}),
+				id: ids.menu,
 				'aria-labelledby': ids.trigger,
 			};
 		}
@@ -94,6 +99,10 @@ export function createSelect(args?: CreateSelectArgs) {
 
 				return isOpen;
 			});
+		});
+
+		attach('mousedown', (e) => {
+			e.preventDefault();
 		});
 
 		return {
@@ -116,31 +125,31 @@ export function createSelect(args?: CreateSelectArgs) {
 		}),
 	}));
 
-	type OptionArgs = {
-		value: string;
-	};
+	type OptionArgs =
+		| {
+				value: string | number;
+		  }
+		| string
+		| number;
 
 	const option = elementMultiDerived([selected], ([$selected], { attach }) => {
-		return ({ value }: OptionArgs) => {
-			attach('click', (e) => {
-				const el = e.currentTarget as HTMLElement;
+		return (args: OptionArgs) => {
+			const value = typeof args === 'object' ? args.value : args;
+
+			attach('click', () => {
 				selected.set(value);
-				selectedText.set(el.innerText);
 				open.set(false);
 			});
 
 			attach('keydown', (e) => {
 				if (e.key === kbd.ENTER || e.key === kbd.SPACE) {
-					e.stopPropagation();
-					e.stopImmediatePropagation();
-					const el = e.currentTarget as HTMLElement;
+					e.preventDefault();
 					selected.set(value);
-					selectedText.set(el.innerText);
 					open.set(false);
 				}
 			});
 
-			attach('mouseover', (e) => {
+			attach('mousemove', (e) => {
 				const el = e.currentTarget as HTMLElement;
 				el.focus();
 			});
@@ -154,6 +163,8 @@ export function createSelect(args?: CreateSelectArgs) {
 				role: 'option',
 				'aria-selected': $selected === value,
 				'data-selected': $selected === value ? '' : undefined,
+				'data-value': value,
+				'data-type': typeof value,
 				tabindex: 0,
 			};
 		};
@@ -239,12 +250,52 @@ export function createSelect(args?: CreateSelectArgs) {
 		}
 	});
 
-	const isSelected = derived(
-		[selected],
-		([$selected]) =>
-			(value: string) =>
-				$selected === value
-	);
+	effect([menu], ([$menu]) => {
+		tick().then(() => {
+			const menuEl = getElementByMeltId($menu['data-melt-id']);
+			if (!menuEl) return;
 
-	return { trigger, menu, open, option, selected, selectedText, arrow, isSelected, options };
+			const selectedOption = menuEl.querySelector('[role=option][data-selected]') as
+				| HTMLElement
+				| undefined;
+			if (selectedOption) {
+				const data = selectedOption.getAttribute('data-value');
+				if (data) {
+					if (selectedOption.getAttribute('data-type') === 'number') {
+						selectedText.set(+data);
+					} else {
+						selectedText.set(data);
+					}
+				}
+			}
+		});
+	});
+
+	const isSelected = derived([selected], ([$selected]) => {
+		return (value: string | number) => {
+			return $selected === value;
+		};
+	});
+
+	const input = derived([selected, options], ([$selected, $options]) => {
+		return {
+			type: 'hidden',
+			name: $options.name,
+			value: $selected,
+			'aria-hidden': true,
+			hidden: true,
+			tabIndex: -1,
+			required: $options.required,
+			disabled: $options.disabled,
+			style: styleToString({
+				position: 'absolute',
+				opacity: 0,
+				'pointer-events': 'none',
+				margin: 0,
+				transform: 'translateX(-100%)',
+			}),
+		};
+	});
+
+	return { trigger, menu, open, option, selected, selectedText, arrow, isSelected, options, input };
 }