+++ src/routes/docs/builders/switch/preview.svelte
@@ -11,8 +11,7 @@
 			<label class="pr-4 leading-none text-white" for="airplane-mode"> Airplane mode </label>
 			<button
 				{...$root}
-				class="relative h-6 w-11 cursor-default rounded-full bg-magnum-800 outline-none
-				transition-colors focus:ring focus:ring-magnum-400 data-[state=checked]:bg-magnum-950"
+				class="relative h-6 w-11 cursor-default rounded-full bg-magnum-800 transition-colors data-[state=checked]:bg-magnum-950"
 				id="airplane-mode"
 			>
 				<span