+++ src/routes/docs/builders/tooltip/preview.svelte
@@ -34,7 +34,7 @@
 <style lang="postcss">
 	.trigger {
 		@apply inline-flex h-9 w-9 items-center justify-center rounded-full bg-white p-0 text-sm font-medium;
-		@apply text-magnum-900 transition-colors hover:bg-white/90 focus-visible:outline-none;
+		@apply text-magnum-900 transition-colors hover:bg-white/90;
 		@apply focus-visible:ring focus-visible:ring-magnum-400 focus-visible:ring-offset-2;
 	}
 </style>